CHAT_BOT_WEB/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.0
Werkzeug==3.0.1
